brushed

+++ dj.py
@@ -8,8 +8,8 @@ client = commands.Bot(command_prefix='>',description="aaja baby aaja tera gana b
 client.remove_command('help')
 
 
-station=705376208745005121
-dj = 707218152307818507
+station=705376208745005121 #channel id
+dj = 707218152307818507 #bot member id
 
 
 
@@ -40,7 +40,9 @@ async def help(ctx):
 
     >pause                                  *to pause*
     
-    >resume                                 *to resume*      
+    >resume                                 *to resume*    
+
+    >vol *1 - 100*                          *volume*
 
     '''
     embed.set_author(name='help')
@@ -52,10 +54,11 @@ async def help(ctx):
 async def join(ctx):
     channel = client.get_channel(station)
     ind=0
-    mem = client.get_all_members()
+    mem = channel.members
     
-    print(type(mem),mem)
+    # print(type(mem),mem)
     voice = get(client.voice_clients, guild = ctx.guild)
+    
     mapp = channel.voice_states
     key = mapp.keys()
     for i in mapp:
@@ -63,9 +66,11 @@ async def join(ctx):
             for i in mem:
                 if i.id == dj:
                      await i.move_to(None)
+                     if voice and voice.is_connected():
+                        await voice.disconnect()
             break
     voice = await channel.connect()
-    await ctx.channel.send("connected")
+    await ctx.channel.send("connected 🟢")
 
 
 
@@ -76,29 +81,37 @@ async def dc(ctx):
     for i in vc:
         if i.channel.id == station:
             await i.disconnect()
-            await ctx.channel.send(f"kr diya leave {channel}")
+            await ctx.channel.send(f"kr diya leave {channel} 🔴")
             return
 
     await ctx.channel.send("already dc hu")
 
 @client.command()
-async def play(ctx, url: str):
+async def play(ctx, url= None):
+    if url == None:
+        await ctx.send(f"⚠ Link to dey ⚠")
+        return
+    url = str(url)
+    voice = get(client.voice_clients,guild=ctx.guild)
+
+    if not (voice and voice.is_connected()):
+        await join(ctx)
+
     song = os.path.isfile("song.mp3")
     try:
         if song:
             os.remove("song.mp3")
             print("Removed old song file")
     except PermissionError:
-       
         print("Song is being played")
-        await ctx.send("some song's playing already")
-        await stop(ctx)
+        if voice.is_playing() or voice.is_paused():
+            await stop(ctx)
         os.remove("song.mp3")
         print("Removed old song file")
         
     
-    await ctx.send("getting everything ready")
-    voice = get(client.voice_clients,guild=ctx.guild)
+    await ctx.send(f"getting everything ready ⏳⏳")
+    
 
     ydl_opts = {
         'format' : 'bestaudio/best',
@@ -115,15 +128,23 @@ async def play(ctx, url: str):
     
     for file in os.listdir("./"):
         if file.endswith(".mp3"):
-            name = file
-            os.rename(file, "song.mp3")
-    
+            if os.name != "song.mp3":
+                name = file
+                os.rename(file, "song.mp3")
+            else:
+                os.remove("song.mp3")
+
+    voice = get(client.voice_clients,guild=ctx.guild)
+
     voice.play(discord.FFmpegPCMAudio("song.mp3"),after = lambda e: print(f"{name} has finished playing"))
     voice.source = discord.PCMVolumeTransformer(voice.source)
-    voice.source.volume = 0.8
+    voice.source.volume = 0.20
 
-    nname = name.rsplit("-",2)
-    await ctx.send(f"playing: {nname}")
+    nname = name.rsplit("-",2)[0]##############testing 
+    p = "▶"
+    m = "🎵"
+    
+    await ctx.send(f" {p} **playing**: {nname} {m}{m}{m}")
 
 
 
@@ -135,7 +156,8 @@ async def pause(ctx):
     if voice and voice.is_playing():
         print("Music paused")
         voice.pause()
-        await ctx.send("Music paused")
+        pa = "⏸"
+        await ctx.send(f"{pa} Music paused")
     else:
         print("Music not playing failed pause")
         await ctx.send("Music not playing failed pause")
@@ -147,9 +169,10 @@ async def resume(ctx):
     voice = get(client.voice_clients, guild=ctx.guild)
 
     if voice and voice.is_paused():
+        p = "▶"
         print("Resumed music")
         voice.resume()
-        await ctx.send("Resumed music")
+        await ctx.send(f"{p} Resumed music")
     else:
         print("Music is not paused")
         await ctx.send("Music is not paused")
@@ -161,12 +184,24 @@ async def stop(ctx):
     voice = get(client.voice_clients, guild=ctx.guild)
 
     if voice and voice.is_playing():
+        r="🛑"
         print("Music stopped")
         voice.stop()
-        await ctx.send("Music stopped")
+        await ctx.send(f"{r} Music stopped {r} ")
+    elif voice and voice.is_paused():
+        voice.resume()
+        await stop(ctx)
     else:
         print("No music playing failed to stop")
         await ctx.send("No music playing failed to stop")
 
+@client.command()
+async def vol(ctx,vol):
+    vol = int(vol)
+    vol = vol/100
+    voice = get(client.voice_clients,guild=ctx.guild)
+    voice.source.volume = vol
+    await ctx.send(f"🔊  {(int)(vol*100)}")
 
 client.run(os.environ['token'])#for heruku)
+
